@PostConstruct组件方法不执行，修改了反射扫描所有packet

+++ mmorpg-server/src/main/java/com/tryingpfq/logic/account/handler/AccountHandler.java
@@ -22,9 +22,9 @@ public class AccountHandler {
     public AbstractPacket login(GameSession session, ReqLoginPacket packet){
         RespLoginPacket  respPacket = new RespLoginPacket();
         if(accountService.login(packet.getAccount(),packet.getPsw())){
-            respPacket.setStatus((byte)1);
+            respPacket.setStatus(1);
         }else{
-            respPacket.setStatus((byte)0);
+            respPacket.setStatus(0);
         }
         return respPacket;
     }
